Revert "feat(schemas): modularize Mailchimp schemas and update documentation"
