--- src/schemas/mailchimp/index.ts
@@ -1,7 +1,0 @@
-export * from "./mailchimp-campaigns";
-export * from "./mailchimp-dashboard-pagination";
-export * from "./campaign";
-// Future: export * from "./audience";
-// Future: export * from "./report";
-// Future: export * from "./member";
-// Future: export * from "./template";
